Initial Features
A struct for a heap was added, as well as an insert function (hInsert), a function to return the 'height' of a heap, one to print it as an array (printHeapArray) and one to print it as a tree (printHeap). printLevel was also added in order to assist the process of printHeap, and recursiveHeight to assist heapHeight. main.c got examples of the heap structure implemented